Initial wip Deno refactor
Starting to migrate from Fresh to standard Deno.
