+++ .github/workflows/deploy.yml
@@ -23,12 +23,6 @@ jobs:
         with:
           deno-version: v2.x
 
-      - name: Build step
-        run: "deno task build"
-        env:
-          SHOPIFY_SHOP: ${{ secrets.SHOPIFY_SHOP }}
-          SHOPIFY_ACCESS_TOKEN: ${{ secrets.SHOPIFY_ACCESS_TOKEN }}
-
       - name: Upload to Deno Deploy
         uses: denoland/deployctl@v1
         with:
