--- dev.ts
@@ -1,6 +1,0 @@
-#!/usr/bin/env -S deno run -A --watch=static/,routes/
-
-import dev from "$fresh/dev.ts";
-import "$std/dotenv/load.ts";
-
-await dev(import.meta.url, "./main.ts");
