+++ islands/Cart.tsx
@@ -8,7 +8,7 @@ import {
   formatCurrency,
   removeFromCart,
   useCart,
-} from "@/utils/data.ts";
+} from "../lib/data.ts";
 
 // Lazy load a <dialog> polyfill.
 // @ts-expect-error HTMLDialogElement is not just a type!
